requirement.txt
langchain 
langgraph 
pandas 
matplotlib 
psycopg2-binary 
python-dotenv 
langchain-google-genai